+++ backend/src/error.rs
@@ -2,7 +2,9 @@ use axum::response::{IntoResponse, Response};
 use deadpool_redis::{CreatePoolError, PoolError};
 use redis::RedisError;
 use reqwest::StatusCode;
-use shared::{AddQuestionValidation, PasswordValidation, TagValidation};
+use shared::{
+    AddQuestionValidation, ContextValidation, EditMetaData, PasswordValidation, TagValidation,
+};
 use thiserror::Error;
 
 use crate::{eventsdb, payment::PaymentError, tracking};
@@ -38,6 +40,12 @@ pub enum InternalError {
 
     #[error("Tag Validation")]
     TagValidation(TagValidation),
+
+    #[error("Context Validation")]
+    ContextValidation(ContextValidation),
+
+    #[error("Meta Validation")]
+    MetaValidation(EditMetaData),
 
     #[error("Events DB Error: {0}")]
     EventsDB(#[from] eventsdb::Error),
@@ -110,18 +118,24 @@ impl IntoResponse for InternalError {
                 (StatusCode::BAD_REQUEST, "").into_response()
             }
 
+            Self::MetaValidation(e) => {
+                tracing::warn!("meta validation: {:?}", e);
+                (StatusCode::BAD_REQUEST, "").into_response()
+            }
             Self::AddQuestionValidation(e) => {
                 tracing::warn!("add question validation: {:?}", e);
                 (StatusCode::BAD_REQUEST, "").into_response()
             }
-
             Self::PasswordValidation(e) => {
                 tracing::warn!("password validation: {:?}", e);
                 (StatusCode::BAD_REQUEST, "").into_response()
             }
-
             Self::TagValidation(e) => {
                 tracing::warn!("tag validation: {:?}", e);
+                (StatusCode::BAD_REQUEST, "").into_response()
+            }
+            Self::ContextValidation(e) => {
+                tracing::warn!("context validation: {:?}", e);
                 (StatusCode::BAD_REQUEST, "").into_response()
             }